% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/csvlr_main.R
\name{bind_all}
\alias{bind_all}
\title{bind_all

Loops through all the dataframes in a list, and binds them all based on common variables.}
\usage{
bind_all(db_list, print_dropped = FALSE)
}
\arguments{
\item{df_list:}{List of dataframes(can be obtained from add_to_list function)}

\item{print_dropped:}{{TRUE/FALSE}, whether to print the list of dropped variables. Default is FALSE.}
}
\value{
A dataframe, containing the merged data
}
\description{
bind_all

Loops through all the dataframes in a list, and binds them all based on common variables.
}
\keyword{csv}

