% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/csvlr_main.R
\name{bind_common}
\alias{bind_common}
\title{bind_common

Binds 2 dataframes based on common variables only. Standalone function which can be operated outside of bind_all.}
\usage{
bind_common(dfr1, dfr2, print_dropped = FALSE)
}
\arguments{
\item{dfr1:}{The first dataframe to bind}

\item{dfr2:}{The second dataframe to bind}

\item{print_dropped:}{{TRUE/FALSE}, whether to print the list of dropped variables. Default is FALSE.}
}
\value{
A dataframe, containing the merged data
}
\description{
bind_common

Binds 2 dataframes based on common variables only. Standalone function which can be operated outside of bind_all.
}
\keyword{csv}

