% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/csvlr_main.R
\name{csv_to_list}
\alias{csv_to_list}
\title{csv_to_list

Function Description:
Adds a folder of CSV files and creates a list containing each CSV as an element}
\usage{
csv_to_list(directory, subdir = FALSE)
}
\arguments{
\item{directory:}{A string, file path containing the CSV files}

\item{subdir:}{{TRUE/FALSE}, whether to include files in sub=directories. Default is FALSE}
}
\value{
A list, containing each CSV as an element
}
\description{
csv_to_list

Function Description:
Adds a folder of CSV files and creates a list containing each CSV as an element
}
\keyword{csv}

