% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/csvlr_main.R
\name{merge_csv}
\alias{merge_csv}
\title{merge_csv

Takes multiple CSV files in a folder and merge them into a combined dataframe}
\usage{
merge_csv(directory, replace_NA = FALSE, subdir = FALSE)
}
\arguments{
\item{directory:}{A string, file path containing the CSV files}

\item{replace:}{{TRUE/FALSE}, whether to replace NA with 0s. Default is FALSE}

\item{subdir:}{{TRUE/FALSE}, whether to include files in sub=directories. Default is FALSE}
}
\value{
A single dataframe, containing the merged CSV data
}
\description{
merge_csv

Takes multiple CSV files in a folder and merge them into a combined dataframe
}
\keyword{csv}

